git-mergelocalbr: ENH: Allow specifying one-by-one merging via special -s1 / -s single argument
And adapt the help when the normal merge fails to include both the spelled-out git merge && ... command pipeline, or the invocation of git-mergelocalbr itself with -s1 added. A complication here is that this command is also invoked by git-mergelocalbrc with the special --branch "br1\nbr2\n" arguments, and that must be properly maintained / handled.