git-tag-version: Allow to pass <commit>
So far, the command assumes the VERSION is the last argument, but "git tag" also supports a <commit>|<object> as the last argument. In fact, there's only the -f(orce) as the only other argument (that's not yet used by the wrapper script).
Change the command-line parsing to account for -f and allow for an optional <commit>.