git-writebackup-import: Order returned from "find" is not necessarily sorted
This can lead to a different commit order of backups done on the same day.
I noticed this because some obsolete files weren't deleted by the import, because the import order was b->c->a, so the removal after c was undone by committing a again. In retrospect, that bug may also have accounted for some weird reversed commit diffs that I noticed when reviewing commit messages.