git-shortblame: Don't use pager unless output is to terminal
I use this command as a data source for the Vim vcscommand.vim plugin, where it is invoked by system(). In GVIM, no real terminal is connected and I get "tput: No value for $TERM and no -T specified" from my $PAGER (smartless).
Test for output to terminal and only use the pager then. To avoid a needless piping to "cat" (which also would require conditional handling of the pager argument), but also avoid duplicating the long command, it is put into a local function.