git-remotebr: Ignore error from git-tracks, and drop the ERROR: prefix from the status message
git-tracks prints "fatal: no upstream configured for branch FOO" when the branch does not track anything. As the command does its own reporting, this message is superfluous and should be suppressed.
This command is reused by other commands; as we don't know the exact context (are missing remote branches just noted, or an actual error), better keep the message neutral by dropping the ERROR: prefix.