Refactoring: Factor out git-parse-last-remotebr-argument, and delegate to git-remotebr
Determining the remote branch for the passed branch is done (better, because it also considers tracked branches) by git-remotebr already. We just need a script that looks for a (local) branch in the last command-line argument, and returns the modified arguments.
Replace the mostly duplicated functionality in git-ucatchup and git-uhardreset with it.