FIX: git-ucatchup needs separate "upstream" "BRANCH" arguments instead of "upstream/BRANCH" (after aa546c2a8e74874d4433a9b7953cba33ac87d1e5)
Add --separate-remote-and-branch argument to git-parse-last-remotebr-argument.
Instead of using git-uup in git-ucatchup (and picking apart the command-line arguments), we can use git-up; the last two arguments are already in the right order (and this is slightly more efficient).
We just need to pick the (last) path argument for the push to origin.