git-localbrdo: Change command-line argument handling for COMMAND
The syntax for accepting certain pipe and list operators inline had weaknesses and did not offer a consistent model. Rather, adopt a different model that either handles simple commands, or requires the quoting of an entire command-line, without exceptions.
The storing / recalling of COMMAND needs rework, as it could be either a SIMPLECOMMAND or COMMANDLINE(s). Because the reading of $commandLineStore is easiest to do into a single argument (without word splitting), we need eval() to execute in any case. It's then just a matter of writing $commands in the proper format.