git-rev-range-this-branch: ENH: Add --one-more[-command] to execute a separate command for the commit before the current branch
${previousRef}^ doesn't work well: In case of a merge commit, it will also show all merged commits; distinguishing with ^1 / ^2 also doesn't help. The only way (I found) to get a single additional commit is using --max-count 1.
Also, some aliases should run a different command for the additional commit; for example, the lc command shows both log and changes, but we only want the log for the additional commit. The --one-more-command offers to supply that exceptional command.
Because there are now two separate git commands in sequence for that, we need separate paging. That pipeline causes Git to turn off automatic coloring, so we need to explicitly enable it, and consider an explicit user instruction to turn it off.