Add git-wrapper and use in git-rev-range and git-pickaxe
It is not possible to override built-in Git commands with aliases or "git-builtin" shell commands. However, I use this, and use a special git() function in interactive shells. In order to have custom Git commands also consider those, I need a git-wrapper that does what my shell function does.
This is mostly for git-log, where I add the --decorate option. It is / can be used by git-rev-range and git-pickaxe.