Add bm... log variants that cover the differences of the current branch to master
When one has already updated the original branch, the br... commands won't work any longer, as the pointer to the original branch has moved on. A common use case is branching off of master, and for that we can reuse the rev-range plumbing to create the master..HEAD revision range.