git-lcl: ENH: Limit the default number of commits to 20
It's too easy to forget about this and launch the command without any arguments, and (on most repositories) this will then take a very long time and probably even has to be aborted on large repos. Better default to a small sensible number unless it is overridden by an explicit argument.