Introduce git-commit wrapper
This allows us to perform custom actions before and after committing, without having to define or interfering with Git hooks.
To override the built-in subcommand in aliases, we have to use our git-wrapper.
Adapt existing scripts to invoke git-commit instead of "git commit".
Note: We do not invoke the wrapper for pure amend actions.