Renaming: git-lllurl to git-lllbrurl
With this, it's explicit in the chosen subcommand what permanency (go with the current branch in lllbrurl vs. permanently link to the current hash in lllhurl) is desired. Had I kept the original name, there'd be the danger of preferring this shorter command, but it often is the wrong choice.