git-inside: Change command-line argument parsing to actually support multiple FILE arguments
Cannot separate SIMPLECOMMAND from options with --; this needs to be the exclusive separator to FILE (in order to also support a single FILE after it). That's okay because it's unlikely that a COMMAND starts with -.
Parse the FILE [...] into files and iterate over them.
FIX: $@ (which only contained the single FILE as $1) was mistakenly passed to execution of COMMANDS. They don't take any passed arguments; all args must be directly passed.