Add coh = checkout-head and underlying unapply-index commands, correct cheatsheet
Contrary to what's documented, "checkout HEAD" does not put the last committed version into the working copy while leaving the index as-is; instead, checkout takes from index and just falls back to HEAD.
According to https://stackoverflow.com/questions/13146908/how-to-revert-changes-in-the-working-directory-only there's no simple solution, so add a low-level git-unapply-index (taken from https://stackoverflow.com/a/38846430/813602) and a git-checkout-head (aliased to "coh", analog to "dh") for this use case.
This is beneficial for example when you've added a bug fix and corresponding test, and now want to verify that the test fails without the fix. Without committing, you have to revert the working copy to HEAD, but keep the fix in the index.