FIX: dt is still a duplicate alias for tdd, and both are broken by the recent changes; introduce git-diffuntil to fix this
As "git diff" doesn't understand the --since argument that git-today-command is now passing. So we first need to supply --no-range, but that commit ID falls one commit short.
From git-dp, we can derive a variant that covers the whole range of commits, and use that to fix the tdd alias.