git-{mine,today}-command: FIX: Need to consider nested *-command in argument distribution
For example, in
>     lgminetoday = today-command mine-command lg
the today-command wrapper would invoke mine-command (as $1), but then place the today filter (--since DATE) after it, but mine-command expects the actual Git command (lg) as the first argument, and cannot execute --since.
We could (optionally) indicate the separation between Git command arguments and remainder with --, but that could interfere with the same argument appended by the user.
We could extend git-wrapper to split $1 into words, and then define lgminetoday = today-command 'mine-command lg', but that would only work with simple arguments.
Instead, make the two *-command wrappers (well, we actually right now only need this for git-today-command) to recognize *-command as a special Git command, and then add the next argument (and any --options in between, meant for the second *-command) to the Git command.