git-wips: Add explicit check for local-only repository
I just check out some projects from upstream, without a personal fork; git ooutlgn then prints "origin does not exist as a remote", which isn't interesting. Instead, I just want to be alerted when _no_ remote is defined yet.
Wrap the ooutlgn in a conditional for existence of origin, and add a general check for remotes before it.