git-selected-command: ENH: Also filter passed file(s) throug the selected command instead of letting the user choose from all of them
We now have so many variants (for staged, changed, untracked) files. The passed file(s) might adhere to the command's filter, or not. It's more intuitive to have the numbers correspond only to applicable files (which the user might have queried first, e.g. via git status), and not to the entire file glob, and filtering the interactive selection also makes more sense.