git-wc{-with-suffix-,,s-in-dir-}do: Need to explicitly forward / parse Git color arguments
Otherwise, the -*) glob triggers and git-wc-with-suffix-do and git-wcs-in-dir-do complain about an unknown option.
git-wcdo previously kept any unknown options together with the working copy arguments; now that we expand these to full absolute paths, this doesn't work any longer (resulting in "ERROR: Cannot chdir to /home/inkarkat/work/--color=always!"), and we need to parse them (into the existing colorArg array). A nice side effect of that is that we don't require the contains command any longer, as it's enough to store the _last_ color argument in case multiple are given, and can do a simple comparison now.