git-wrapper: Also handle hub aliases (if hub is installed) and use for delegation instead of hub || git
When custom git-* / hub-* commands are passed to e.g. git-wcdo, the alias needs to be resolved there as well. Make git-wrapper also handle hub aliases (if hub is installed); keep the dedicated hub-wrapper for pure hub wrapping. (As hub falls back to Git built-in commands, hub-wrapper therefore handles hub-aliases, and built-in hub and git commands, but not git-aliases.)