git-wcs-in-dir-do: ENH: Add --skip-symlinks option
As this is hard to check from within the working copy with --predicate-command, and also way clearer.
This is useful when a working copy is symlinked into several directory trees, e.g. I have the runVimTests project once in the .vim config and once as a bin subdirectory for the shell commands.