git-selected-command: BUG: Regression: git-addselectedhunk broken (after 25aabc77be6fabf6505019295f5d73d47b0c7720)
The Git command itself can consist of command + arguments (in the case of git-addselectedhunk: "add --patch"), so it must not be quoted.

The fallback implementation had had this problem from the very beginning (because it quoted "$gitCommand"); it was only the interpolation inside --command "git-wrapper $gitCommand" that implemented that correctly.
Now, the ${gitCommand[@]} array contains the properly separated arguments for both, and --command uses extra quoting to maintain it.