git-pickaxe: Drop the GIT-COMMAND first argument
Using pickaxe on show isn't so useful after all (as a commit has to be specified; it doesn't search for the first commit matching; for that, git log -1 has to be used). Passing the Git command made direct use of the command more awkward. Instead, just allow customization via GIT_PICKAXE_COMMAND.