git-pickaxe: Add --more that uses -G (grep) instead of -S and define touched... variants of find... aliases
Whereas -S needs a difference in the number of appearances / matches in a change (so it observes additions and deletions, but not simple modifications), -G accepts any changed line that matches.

,----[ git help log ]----
| To illustrate the difference between -S<regex> --pickaxe-regex and
| -G<regex>, consider a commit with the following diff in the same
| file:
|
|     +    return !regexec(regexp, two->ptr, 1, &regmatch, 0);
|     ...
|     -    hit = !regexec(regexp, mf2.ptr, 1, &regmatch, 0);
|
| While git log -G"regexec\(regexp" will show this commit, git log
| -S"regexec\(regexp" --pickaxe-regex will not (because the number of
| occurrences of that string did not change).
`----

Unfortunately, in yet another bout of Git command-line interface inconsistency, -S can do both literal and regexp (with --pickaxe-regex) matching, but -G always takes a regexp. Let's iron that out inside our wrapper script, which now supports the grepping with a custom --more argument.

Add touchedlogs variant of findlogs (et al.); the names are long, but so far I haven't used the commands often enough that I would wish for a shorter alias; rather, discoverability of the functionality is important.