Add missing special *others variants of *mine commands: lgothers and lastothers
We don't need a separate script for lgothers; lgmine only does this to modify the lg output format to omit the (single and well-known, as it's me) author from the output. As lgothers is about different (and potentially multiple) people, keep the default format; this can be achieved with a simple alias.