CHG: Reverse the order of *lc* aliases (that operate on a commit range) so that the earliest come first
When reviewing changes done today, or a branch, etc., one usually wants to go in chronological order, from the earliest to the latest ones. As there's no -r short form of --reverse, doing this by hand is too cumbersome (remembering this, too). The only downside of adding this to the alias is that one cannot deliberately reverse back again to the anti-chronological order (this would just pass --reverse --reverse).