git-logv: Default to "fuller" log format that includes separate commit and author fields
So far, the v[erbose] just meant including a list of affected files (as in Subversion). But actually it makes sense to add more commit data (especially the difference of committer vs. author (names and dates)), so that the command can be used for that use case as well (and we don't need to define a separate command (with all the varieties)).