Add commituntracked/ cun
We cannot just pass the untracked files to "git commit", as those arguments must be known to Git already. So there's no way around adding them to the index first. In order to keep the "directly commit, bypassing the index" semantics, any existing changes in the index have to be stashed first.