git-wrapper: Lookup hub / git executable instead of hard-coding git
Some aliases that invoke another alias failed when invoked through git-wrapper (e.g. lhuurl1 -> lhuurl -> git-lhurl), but succeed in an interactive shell. The reason is that the git wrapper function does not directly invoke /usr/bin/git, but uses the hub (itself a wrapper) if it exists.
Consistently use the lookup here as well.