git-wips: Need explicit color argument handling
When this is invoked as part of my dashboard, the output is redirected (for capture by Conky), but the git ooutlg still prints ANSI escape sequences. Can't suppress this via "git wips --no-color", as the dashboard command chain (through git-wcdo) doesn't pass this on. Actual detection in here is more comfortable.