Add --push to git-reintegrate and add hub-reintegrate* overloads
For GitHub's pull request interface, the ordering of actions matters: When the feature branch is deleted in origin first, and only then the merged (master) changes are pushed, GitHub sees the branch deletion first and marks the pull request as "Closed" instead of "Merged". The end result is the same (the branch is merged and the feature is in), but GitHub doesn't recognize it, which isn't nice.
We need to first push the merged branch and only then delete the feature branch (as one would do directly in GitHub's web interface). For that, add --push to git-reintegrate and add hub-reintegrate* overloads that automatically set that option.