git-up: Also do a fetch first if GIT_UP_FETCH is set
As "git pull" will only fetch the remote branch that is tracked by the current one, but won't get in any other (new or updated) remote branch. We may be interested in those branches, too.
For implemenation, though the documentation states that "git pull is shorthand for git fetch followed by git merge FETCH_HEAD", it doesn't seem to be so easy to emulate together with --rebase. "git rebase" doesn't correctly understands FETCH_HEAD (and tries to rebase on the first branch from in there instead of picking the right one).
Therefore, just do a separate fetch first; this will update the current remote branch twice, but the remote will always correspond to the local because we pull after the fetch.