CHG: Rename *oth to *lastby and *others to *by
The ...by suggests that this command needs author(s) as a parameter (either passed or else queried).
The ...lastby indicates a further restriction on the last consecutive range by that author.
I hope that these are easier to memoize; at least ..by is much shorter than ...others, and we don't need a hack like ...oth (I think the shortening should indicate that this also would result in a shorter list... which was not terribly intuitive).