git-reintegrate: FIX: --rebase does not work
As we want to rebase the branch on the current branch (e.g. master), need to pass the <upstream>, i.e. HEAD. The rebase command checks out the branch, resets it to <upstream>, and applies the changes. The following merge still needs to be on the <upstream> branch, so first record the current ref (e.g. master), and then restore it.