git-orig-ref: Fall back to returning SHA1 hash (in case of detached working copy), force symbolic resolution with new --symbolic
All of my clients so far only use the ref to be able to return to it (via git checkout "$(git-orig-ref)"); that easily also can handle a detached working copy by returning the SHA1 hash. Any clients that would require an actual symbolic ref (and need to fail on a detached working copy) can request that via --symbolic.