Store base branch in $GIT_DIR instead of just an ephemeral environment variable GIT_BASE_BRANCH
This way, it'll persist across shells; the designation of a base branch usually is something long-term. We need to use a file in $GIT_DIR instead of a local Git configuration to have it scoped to the worktree (local Git config is shared across all worktrees).
Define a new git-nbr accessor and adapt all existing commands to switch from GIT_BASE_BRANCH to that.
Add git-nclear subcommand to clear the (now persistent) base branch.
git-nco becomes a plain subcommand instead of the shell alias, as it does not need to modify the current shell's environment any longer (another benefit).