git-reintegrate, git-rebaselocalbr: Drop --preserve-merges
This only sounds useful when pulling changes (in git-up, and even there only in rare corner cases, which are documented there now), but not when actually rebasing a branch (where the main intention is to keep a single linear stream of changes).
Oh, and --preserve-merges has been deprecated in Git 2.22 (in favor of a new --rebase-merges), anyway.