git-{dirty,untracked}: Handle passed <pathspec>
For git-untracked, this is a simple matter of passing through the arguments to git ls-files.
For git-dirty, this would only work for git diff-index, but even though git update-index support <file> arguments, --verbose shows that it would then consider other files as to be added. So we have to switch algorithms then and check for output of git status while excluding untracked files.