git-up: FIX: Rebase actually needs to be passed upstream/master, not upstream master
As that only results in
> fatal: Needed a single revision
> invalid upstream upstream
(The git rebase documentation is really hard to read.)

We already have the latter format in $path, which is derived from the last passed argument (so all except the last need to be forwarded).
Expand gitArgs into a conditional and now do the separate calls for rebase or pull.