git-rebaselocalbr: ENH: Allow pushing to tracked / passed branch via --push / --push-to
Every successfully rebased branch is pushed, so that when a failure occurs, the rebasing can be restarted and no pushes are lost.
If the push itself fails, iteration doesn't abort (as this is side functionality and the push can easily later be done, even in bulk), but the exit status changes.