git-pickaxegrep: Forward passed --max-count and --reverse arguments directly to git-logrevs
This simplifies the where{last,introduced}l*g commands because we don't need to use the GIT_PICKAXEGREP_LOGARG variable to channel the arguments to the right position.
Also, this restores the correct behavior from the user perspective, where --max-count always affects the number of elements that are output, not the number that are searched. In other words, git-pickaxegrep now is consistent with git-pickaxe and git-logg. However, we cannot limit the number of searched commits any longer (which could be important as the entire history is searched, and that could take considerable time in large repos). Let's introduce a separate --source-count for that.