git-{archive,,merge,rebase}localbr{,do,refdo}: FIX: Consistently parse --branches argument with proper handling of IFS
Do the splitting of the (usually space-separated) "<branch1> ..." argument at the beginning, with the original IFS, then change IFS to \n as that is required to consume the returned lines of branches into an array.
That variable is now consistently an array named localBranches. This also simplifies the linewise output, e.g. in archivelocalbr.