git-whois: Accept single matching person without prompting via commandOnSelected
To be consistent with the fallback implementation.
The situation here is different than with git-selected-command, where despite only one available option, we still want the user to actively confirm that this actually is needed. No such thing is needed with persons: if we have a single match, that's the guy!