Place gitColorArg just behind the Git subcommand instead of simply appending it
Instead of parsing SIMPLECOMMAND into simpleCommand, do this only temporarily into uncoloredSimpleCommand and then restore $@.
Turn gitCommand into an array and move any arguments in front of the Git subcommand to it, then append gitColorArg. We only need to do this if there actually is a gitColorArg. The algorithm for parsing the arguments is taken from the various git-*-command metaprograms.