git-mbr: ENH: Check for actual existence of master or main branches, and use the first that exists
And also persist that information. Without that, the command would continuously print the "Note: ..." if for example no remotes are defined so the above determination would remain inconclusive.
If none of the candidate branches actually exists, print nothing and fail (so clients will notice and not act on wrong information).