hub-crstats: ENH: Allow prepending CR date(s) with --with-{oldest,newest}-date
In Epoch format with prefixed @ (for parsing with reldate); we need the numeric format for the earlier-later comparison. For output, it would be nice to do this as (absolute / relative) dates, and we could query those, but the parsing of the relative one would be difficult (as we're splitting on spaces).