Refactoring: Pass the --filter '' from those git-selected-command clients that want it
It's not just commitselected where the application of the generator makes sense also for any passed files; clients are about evenly split. Therefore, it makes sense to have the client decide what is wanted, and explicitly pass the --filter '' if selection should be done from all files, not just those generated.