git-modified: Also handle staged files where the modification has been undone in the working copy
"git diff --name-only HEAD" compares the working copy with the last commit, so if the staged change is reverted in the working copy, those files are not listed. As this is a rare corner case (but still somewhat important when the list of files powers selections from modified files), and "git status --porcelain --untracked-files=no" apparently internally invokes two diff commands as well (and then requires a cut command to get rid of the prefixed status), let's detect the case via an associative array for the staged files and an array for modifications with regard to HEAD, assuming that there likely will be few modifications, and only perform a merge (involving sort to maintain the correct ordering) if necessary.

Source: https://stackoverflow.com/questions/39994736/git-show-all-modified-files-staged-and-not-staged