CHG: Rename l*gstf to l*gst
I have logi for staged files, am going to add logI for modified but not staged files, so what do I choose for any modified (staged and unstaged) files? I've considered logiI, or loga (all), or logm (modified), but these all aren't very nice. So how about just shortening the overly long logstf (that I've never used) to logst, with the "st" suffix meaning "and file for which there is a status"?!