git-selectedcommit-command: Also recognize --follow as an additional range argument
So that e.g. "git correcti --follow HEAD~100" works; previously, this would cause the second argument (HEAD~100) to be passed on to "git correct", leading to "error: pathspec 'HEAD~100' did not match any file(s) known to git."
Use a separate rangeModificationArgs array instead of appending to rangeArgs, because the latter is checked for the mandatory supplied range arg.