git-byuserrange-command: ENH: Allow control for passing of GIT-COMMAND arguments to the revision range determination and add msg[last]... variants of by...
Clients of git-byuserrange-command are either variants of "git log", or pickaxe extensions that take an additional last text|/regexp/ argument, or completely different. Arguments for the last would confuse the "git log" done to determine the revision range, so introduce additional options --with-log-args and --with-log-args-except-last for the first two use cases, and go over all by... aliases to supply the correct customization.
Additional arguments can both be log options (--max-count, --follow), or -- <path> limits. The tricky thing is that the difference detection also passes the starting revision, and according to git log's usage, we'd strictly would have to split the arguments so that the revision fits right in the middle. Fortunately, git log is more flexible and also handles log options after the <revision>.
Unfortunately, that still isn't enough to support "-- <path>", as the algorithm currently assumes a complete sequential log without skipped commits, and uses the first differing line as an offset from the starting revision. By switching to a custom diff output, we can instead obtain the first different revision (and the count of the first difference for the corner case), and can use that directly (and just append ^ for --one-more). Though still tricky (and with a loss of proper error handling if diff invocation fails), at least we get rid of the ugly parsing of cmp's output.
Apply the same changes to git-bylogmessagerange-command, too, and introduce msg... and msglast... (that have the --auto-revision enabled) variants of by...