CHG: Renaming: lhurl to lurl
The "h" for hash is a relict of its creation: After a log of commit hashes, it was natural to embellish these hashes to the GitHub URLs (which of course have the hash in them), but the actual subject is the (origin|upstream) URL. Let's get rid of the "h".
Bonus: Correct the inconsistency of GITLHURL_... vs. GIT_LHURL_...