git-wcdo: Silence errors from progressNotification
For instance, progressNotificationToAddendum invokdes titleupdate which requires a $TERMID value. Unfortunately, even when it fails, due to the redirection, we don't get a non-zero exit status out of it (just the error message on stderr). Only later, when we attempt to write the first progress in reportProgress() will it fail with a broken pipe, and the script exits with 142 (i.e. 128 + SIGPIPE), which is not what we want - progress reporting is a subordinate concern, and the working copy iteration should proceed.
We use a trap to suppress the SIGPIPE, and in the exception handler redefine the reportProgress() function to a no-op (so this error does not unnecessarily happen on each update). After that, we can restore the trap. (I'm not sure whether keeping it could negatively affect the COMMANDs supplied by the user.)