git-autostash: Documentation: Mention index behavior
And why autostashing will always process both working copy and index.
I was looking at the separate index + working copy change stashing done by git-mergeamend and wondered whether this should be put into git-autostash as well.
I think Git does not separately restore those (where it offers autostashing, e.g. with rebase and pull), because autostashing is needed to avoid potential conflicts with the executed operation, and that means that there may be conflicts during git stash that then need to be manually resolved. During this resolution, the index is needed as a separate store. So according to the Git developers (I assume), it's not worth the effort to save and restore staged and unstaged changes separately, because those would only survive in the happy case.
I think I'll agree here. In any case, it's not recommended to trigger such operations when there's so much turmoil inside the working copy. And I can use my "git stashI" and "git stash" to separately stash both (and then restore with "git stash pop --index" and "git stash pop").