git-stashi: FIX: Check for staged and unstaged changes
If nothing's in the index, we can fail early - the following stashing of the index would fail, anyway, and we'd have to needlessly undo the stashing of the working copy.
If there are no unstaged changes, a plain stash will do (the --no-keep-index is just for documentation; that's the default stash behavior). We can skip the whole temporary stashing of the working copy (and the existing implementation would fail, anyway).