git-reintegrate: Only push to writable remotes
Especially in open source developments, the upstream remote is configured but not directly writable, and updates have to happen through pull requests from origin. git-reintegrate was written in a corporate background, and causes errors when a repository is read-only.
Include the new git-remote-isreadonly in the conditionals to skip pushing when there's no write access.