FIX: Remove --one-more from the br..l[o]gg aliases
The lgg grepping applies to the one-more commit, too; this should only provide context, and therefore that doesn't make sense.
The logg has just been changed to greplog (so without grepping). As the grepping breaks the continuity of commits, there's also no use in the context, anyway, so drop that as well.