git-stashselected: Compatibility: Use "git stash push -- <pathspec> [...]" if available
With this, we can directly stash unstaged changes, without having to go through the index. Legacy Git (< 2.13.7) still falls back to the existing implementation (which now also uses "git-unstaged" as the select source and appends a -- separator for robustness). As the new "git stash push" supports --patch, but not --edit, the latter if given also employs the fallback implementation. (Though granted I've never used that option myself; it was only added for completeness.)