Reimplementation: Use update-ref instead of reset --hard
With update-ref, a non-current branch pointer can be moved; this is exactly what we need: After moving the current commit(s) to the new branch, reset the previous (original) branch.
This reverses the order: We can first create the new branch off the current / unmerged previous commit, and then (from that new branch) drop those commits off the original branch.
So far, we had to first drop the commits off the current branch, and then create the branch based on the memoized original commits. This unnecessarily updated the working copy, and so we had to ensure a clean one.

Source: https://stackoverflow.com/a/1591255/813602