Renaming: sdiff to diffy
The "s" prefix is inconsistent and therefore hard to recall. It was inspired by the "sdiff" tool in GNU diffutils, but even it now defers the side-by-side functionality to "diff -y".
Move the difftool setup from my external ~/.gitconfig in here, as it belongs here. The profile/aliases.sh does similar programmatic difftool (and mergetool) setups, but for diff we can safely assume that it's there.