git-contained-command: BUG: Fallback may get activated by head cutting off tag listing
When there are many tags, the "git tag --contains" enumeration isn't done yet, and the pipeline aborts with exit status 141; unfortunately, that failure activates the fallback, which is printed in addition, resulting in two versions.
Split the two conditional branches, and do the head internally by reading into an array.
ENH: Surround the fallback value with (...) to indicate that it's not an actual version but the fallback.