git-contained-command: Also parse --exclude commands and pass to "git describe"
There's a fundamental difference between gitCommand=describe and gitCommand=annotated: The former identifies a best-match tag (and then includes commit distance, e.g. "v7.1~97"), whereas the latter lists all tags the commit is contained in.
The grepping of tags is suited for the latter, e.g. to filter out intermediate, non-released tags, but it doesn't work with the former, as the single tag would be ignored and the (branch name or SHA) fallback would be printed. Instead, "git describe" has it's own built-in filtering via GLOBs passed via --exclude.
Let's parse those (and append to gitCommand, hoping it'll be "describe" as only that understands it).
And as this topic is quite complex, enhance the usage help to give more precise hints (though I likely will use this solely through my aliases).