ENH: Allow passing label names to "hub labels set"
This significantly limits the number of web service calls when the user knows what has changed (e.g. one label updated, one added). This is an important speedup for global changes to labels across dozens of repositories.
If a label name is passed that does exist in GitHub, but is missing from the label file, we can deduct that this should be removed and do that as well (by delegating to deleteLabels()).