Usability: Limit number of incoming / outgoing / behind commits
By truncating in the middle. A very long list of commits (e.g. when not updating a busy upstream for a long time) is not helpful here (also because git-wips does not have paging built-in), and if necessary, the appropriate commands (git uinlg etc.) can be used to get the full list.