git-up: ENH: Also take the last branch from the reflog for determining the remote branch to log
If we're just following an upstream repository and checking out tagged fixed versions, there's no local branch, and it's not possible to follow a remote branch for changes.
However, we likely had an upstream branch (stable or master) checked out before switching to tagged versions. Try to locate that and use it, to avoid output like this:

> ERROR: Checked out tag 0.29.0
> Warning: Could not determine corresponding upstream branch, no logging possible.