git-wips: ENH: Add behindtags check
When a tagged version is checked out in an original repo, no upstream remote branch can be found, so the "behind" check will not notify of commits that have been added. (We're likely only interested in newer tagged versions, anyway.)
For the new check, iterate all upstream branches (git-brrefdo with git-ubr as the branch source) until (excluding) the current tag is found. As the A..B notation only ensures that A and B are valid refs, but does not ensure that A is a predecessor of B, we need to ensure that the tag actually is reachable on that particular branch, to avoid that other previous tags are mistakenly picked up. A new git-containstag command provides a suitable interface for that.
As multiple branches may find intermediate tags, present a unique list. If that is not empty, we have newer tag(s) to offer to the user to update to, and the check succeeds (due to outputAndPrintf).