ENH: Add unmergeddownstream check
This adds more precision to the downstreambranches check (and therefore goes before it). A branch in origin that is not in upstream likely is an open pull request, but it could also be a left-over merged branch that has long been incorporated (and the PR closed), but was forgotten to be deleted.