Reimplementation: Shorter submodule diff format
Let's make a virtue out of the necessity of coloring the "git submodule summary" output to look like "git diff --submodule=log" and completely restructure the information to be even shorter. After all, "git diff --submodule=log" will provide that information already (and for me is handily configured as "git d").
Instead, opt for an even shorter visualization, taking just the individual submodule headers, and turn every > / < following log line into a diffstat-like display.
Git's mentioning of the overall in/out direction (the "(rewind)") can be used to determine the overall coloring of the submodule status.