ENH: Insert the submodule path into the diff headers
Otherwise, it's hard to recognize the affected files as the submodule information effectively is lost. I don't want to let "git submodule" introduce it by dropping the --quiet, as that's ugly and not at the expected position.
Add a postprocessing step to tweak the diff headers based on the special $displaypath variable set by "git submodule foreach".
As that has to happen on each iteration, do a recursive script invocation with the special --diffmassage flag, to avoid messy escaping and overly long command-lines.