git-localbrdo: Reimplementation: Just list branches when no command; drop the confirmation and --yes, use external store only when interrupted
To interactively iterate over all local branches, it's better and more consistent with the other *do iterations to open a new shell (by passing the "shell" SIMPLECOMMAND) than returning to the invoking shell and relying on --continue. (Other iterations cannot use that approach as they'd need to change directory of the invoking shell - the branch change within the working copy is a special case here.)
With that, the iteration can happen in the localbrdo() loop, with the local branches simply stored in an array. However, the interruption on error and refusal to continue with the working copy is still dirty are nice features that should be kept. I could open a shell in these circumstances (even when the passed COMMAND is not an interactive shell), but I think it's better to then drop back to the original shell. So the persistence mechanism needs to stay, but now is on-demand only. Persistence is to a single file, and just stores the internal variables (like the localBranches array, and all parsed command-line options), with some special considerations to allow a different COMMAND to be passed.